% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipseq_related_functions.R
\name{clipseq_annotate_peaks}
\alias{clipseq_annotate_peaks}
\title{Annotate clipseq peaks}
\usage{
clipseq_annotate_peaks(
  x,
  ref_cds_regions,
  ref_utr3_regions,
  ref_utr5_regions,
  ref_intron_regions.
)
}
\arguments{
\item{x}{an object of class GRanges where each row denotes clipseq peaks. Object must have column 'peak_id'.}
}
\value{
an object of class GRanges
}
\description{
Annotate clipseq peaks
}
\examples{

peak_file <- system.file("extdata", "boudreau_et_al_clipseq_peaks.bed", package = "SpongeAnalysis")
peak_data <- readr::read_delim(peak_file,delim = "\t")
peak_data_gr <- peak_data \%>\% plyranges::as_granges()
clipseq_annotate_peaks(peak_data_gr)

}
