% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipseq_related_functions.R
\name{.annotate_introns}
\alias{.annotate_introns}
\title{categorize introns into introns (coding + non-coding), intron_utr3 and intron_utr5}
\usage{
.annotate_introns(intron_file, utr3_file, utr5_file, coding_exon_file)
}
\arguments{
\item{intron_file}{character string pointing intron regions in a bed file format}
}
\value{

}
\description{
categorize introns into introns (coding + non-coding), intron_utr3 and intron_utr5
}
\examples{
intron_bed_file <- system.file("annot_grch38_genecode_v36", "grch38_genecode_v36_introns.bed", package = "SpongeAnalysis")
utr3_bed_file <- system.file("annot_grch38_genecode_v36", "grch38_genecode_v36_3UTR.bed", package = "SpongeAnalysis")
utr5_bed_file <- system.file("annot_grch38_genecode_v36", "grch38_genecode_v36_5UTR.bed", package = "SpongeAnalysis")
cds_bed_file <- system.file("annot_grch38_genecode_v36", "grch38_genecode_v36_coding_exons.bed", package = "SpongeAnalysis")
.annotate_introns(intron_file = intron_bed_file,utr3_file = utr3_bed_file ,utr5_file = utr5_bed_file,coding_exon_file = cds_bed_file)
}
\keyword{internal}
