% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IR_related_functions.R
\name{.assign_class_spongeAnalysis}
\alias{.assign_class_spongeAnalysis}
\title{Assign class spongeAnalysis}
\usage{
.assign_class_spongeAnalysis(x)
}
\arguments{
\item{x}{a list or dataframe to which class spongeAnalysis to assign.}
}
\value{
an object of class spongeAnalysis.
}
\description{
This function does all mandatory checks before it assigns class spongeAnalysis.
}
\keyword{internal}
