% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipseq_related_functions.R
\name{.get_ucsc_coding_exon}
\alias{.get_ucsc_coding_exon}
\title{Get annotation of coding exons}
\usage{
.get_ucsc_coding_exon(coding_exon_file)
}
\arguments{
\item{coding_exon_file}{a character string pointing coding regions in a bed file format}
}
\value{
an object of GRanges
}
\description{
Get annotation of coding exons
}
\examples{

cds_bed_file <- system.file("annot_grch38_genecode_v36", "grch38_genecode_v36_coding_exons.bed", package = "SpongeAnalysis")
.get_ucsc_coding_exon(cds_bed_file)

}
\keyword{internal}
