% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipseq_related_functions.R
\name{.get_ucsc_introns}
\alias{.get_ucsc_introns}
\title{Get annotation of intron regions}
\usage{
.get_ucsc_introns(intron_file)
}
\arguments{
\item{intron_file}{character string pointing intron regions in a bed file format}
}
\value{

}
\description{
Get annotation of intron regions
}
\examples{
intron_bed_file <- system.file("annot_grch38_genecode_v36", "grch38_genecode_v36_introns.bed", package = "SpongeAnalysis")
.get_ucsc_introns(intron_bed_file)

}
\keyword{internal}
