% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipseq_related_functions.R
\name{.get_ucsc_utr3}
\alias{.get_ucsc_utr3}
\title{Get annotation of utr3 regions}
\usage{
.get_ucsc_utr3(utr3_file)
}
\arguments{
\item{utr3_file}{a character string pointing utr3 regions in a bed file format}
}
\value{

}
\description{
Get annotation of utr3 regions
}
\examples{
utr3_bed_file <- system.file("annot_grch38_genecode_v36", "grch38_genecode_v36_3UTR.bed", package = "SpongeAnalysis")
.get_ucsc_utr3(utr3_bed_file)

}
\keyword{internal}
