% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipseq_related_functions.R
\name{.get_ucsc_utr5}
\alias{.get_ucsc_utr5}
\title{Get annotation of utr5 regions}
\usage{
.get_ucsc_utr5(utr5_file)
}
\arguments{
\item{utr5_file}{a character string pointing utr5 regions in a bed file format}
}
\value{

}
\description{
Get annotation of utr5 regions
}
\examples{

utr5_bed_file <- system.file("annot_grch38_genecode_v36", "grch38_genecode_v36_5UTR.bed", package = "SpongeAnalysis")
.get_ucsc_utr5(utr5_bed_file)

}
\keyword{internal}
