% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IR_related_functions.R
\name{.validate_irfinders_object}
\alias{.validate_irfinders_object}
\title{Check if the object belongs to class spongeAnalysis.}
\usage{
.validate_irfinders_object(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
Check if the object belongs to class spongeAnalysis.
}
\keyword{internal}
