% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IR_related_functions.R
\name{get_intron_master_list}
\alias{get_intron_master_list}
\title{Prepare a master list of introns}
\usage{
get_intron_master_list(x, add_meta_data = T, bs_genome_object = NULL)
}
\arguments{
\item{x}{an object of class spongeAnalysis}

\item{add_meta_data}{logical, whether to map meta data or not}

\item{bs_genome_object}{an object of class BSgenome}
}
\value{
a dataframe
}
\description{
Prepare a master list of introns
}
\examples{
example_dir <- "~/Documents/Projects/15_SpongeAnalysisRpkg/SpongeAnalysis/inst/extdata"
example_files <- fs::dir_ls(example_dir, regexp = "exp*")
names(example_files) <- c("exp1" , "exp2")
x <- read_irfinderS_output(files = example_files[1],  add_prefix_chr = T)
get_intron_master_list(x,bs_genome_object = BSgenome.Hsapiens.UCSC.hg38::BSgenome.Hsapiens.UCSC.hg38)
}
