% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IR_related_functions.R
\name{get_ir_counts}
\alias{get_ir_counts}
\title{Prepare a intron reads count matrix.}
\usage{
get_ir_counts(x)
}
\arguments{
\item{x}{an object of class SpongeAnalysis.}
}
\value{
a tbl.
}
\description{
Intron counts will be obtained from the column 'introndepth' of the IRFinder-S output.
}
\examples{

example_dir <- "~/Documents/Projects/15_SpongeAnalysisRpkg/SpongeAnalysis/inst/extdata"
example_files <- fs::dir_ls(example_dir, glob = "*IRFinder-IR-nondir*.txt")
names(example_files) <- c("exp1" , "exp2")
x <- read_irfinderS_output(files = example_files,  add_prefix_chr = F)
get_ir_counts(x)


}
