% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IR_related_functions.R
\name{map_intron_meta_data}
\alias{map_intron_meta_data}
\title{Map metadata irfinder-s output}
\usage{
map_intron_meta_data(x, bs_genome_object = BSgenome.Hsapiens.UCSC.hg38)
}
\arguments{
\item{x}{an object of class spongeAnalysis.}

\item{bs_genome_object}{an object of class BSgenome}
}
\value{
a dataframe
}
\description{
This function allows mapping DNA sequence, GC content and intron length to irfinder-s output
}
\examples{
example_dir <- "~/Documents/Projects/15_SpongeAnalysisRpkg/SpongeAnalysis/inst/extdata"
example_files <- fs::dir_ls(example_dir, glob = "*IRFinder-IR-nondir*.txt")
names(example_files) <- c("exp1" , "exp2")
x <- read_irfinderS_output(files = example_files[1],  add_prefix_chr = T)
map_intron_meta_data(x = x,  bs_genome_object = BSgenome.Hsapiens.UCSC.hg38::BSgenome.Hsapiens.UCSC.hg38)
}
