% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IR_related_functions.R
\name{mark_ir_status_by_filters}
\alias{mark_ir_status_by_filters}
\title{based on each filters applied it checks whether each intron is retained or not}
\usage{
mark_ir_status_by_filters(
  x,
  min_intron_cov = 0.95,
  min_intron_depth = 5,
  minimum_splice_exact = 5,
  min_irratio = 1e-04
)
}
\arguments{
\item{x}{an object of class sponge analysis}

\item{min_intron_cov}{a numeric value between 0 and 1, default 0.95, denoting minimum value for coverage cutoff}

\item{min_intron_depth}{a numeric value, default 5, denoting average sequence depth for each intron.}

\item{minimum_splice_exact}{a numeric value, default 5, denoting number of reads supporting intron splicing.}

\item{min_irratio}{a numeric value, default 0.0001, denoting ir-ratio.}
}
\value{
an object of class spongeAnalysis
}
\description{
based on each filters applied it checks whether each intron is retained or not
}
\examples{
example_dir <- "~/Documents/Projects/15_SpongeAnalysisRpkg/SpongeAnalysis/inst/extdata"
example_files <- fs::dir_ls(example_dir, glob = "*IRFinder-IR-nondir*.txt")
names(example_files) <- c("exp1" , "exp2")
x <- read_irfinderS_output(files = example_files,  add_prefix_chr = T)
mark_ir_status_by_filters(x)

}
