% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IR_related_functions.R
\name{read_irfinderS_output}
\alias{read_irfinderS_output}
\title{Read IR-finderS output}
\usage{
read_irfinderS_output(files, add_prefix_chr = TRUE, remove_prefix_chr = FALSE)
}
\arguments{
\item{files}{a character vector denoting irfinderS output file(s) ending with suffix "IR-nondir".}

\item{add_prefix_chr}{logical, whether to add prefix 'chr' in the column seqnames}

\item{remove_prefix_chr}{logical, whether to remove prefix 'chr' from the column seqnames}

\item{select_columns}{logical, whether to subset columns. If TRUE below columns will be subsetted
+ chr
+ start
+ end
+ name
+ null
+ strand
+ coverage
+ introndepth
+ spliceleft
+ spliceright
+ spliceexact
+ irratio
+ warnings}
}
\value{

}
\description{
Read IR-finderS output
}
\examples{
example_dir <- "~/Documents/Projects/15_SpongeAnalysisRpkg/SpongeAnalysis/inst/extdata"
example_files <- fs::dir_ls(example_dir, regexp = "exp*")
names(example_files) <- c("exp1" , "exp2")
read_irfinderS_output(files = example_files,  add_prefix_chr = F)
}
