% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IR_related_functions.R
\name{select_cols_irfinderS_output}
\alias{select_cols_irfinderS_output}
\title{Subset columns from IRfinder-S output}
\usage{
select_cols_irfinderS_output(
  x,
  keep_columns = c("chr", "start", "end", "name", "null", "strand", "coverage",
    "introndepth", "spliceleft", "spliceright", "spliceexact", "irratio", "warnings")
)
}
\arguments{
\item{x}{an object of class spongeAnalysis}

\item{keep_columns}{a character vector denoting column names to keep in the output dataframe
+ chr
+ start
+ end
+ name
+ null
+ strand
+ coverage
+ introndepth
+ spliceleft
+ spliceright
+ spliceexact
+ irratio
+ warnings}
}
\value{

}
\description{
Subset columns from IRfinder-S output
}
\examples{
example_dir <- "~/Documents/Projects/15_SpongeAnalysisRpkg/SpongeAnalysis/inst/extdata"
example_files <- fs::dir_ls(example_dir, regexp = "exp*")
names(example_files) <- c("exp1" , "exp2")
x <- read_irfinderS_output(files = example_files,  add_prefix_chr = F)
select_cols_irfinderS_output(x)
}
