% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IR_related_functions.R
\name{sponge_analysis_assign_intron_identifier}
\alias{sponge_analysis_assign_intron_identifier}
\title{For each intron in the object spongeAnalysis assign unique intron id}
\usage{
sponge_analysis_assign_intron_identifier(x)
}
\arguments{
\item{x}{an object of class spongeAnalysis.}
}
\value{
an object of class spongeAnalysis.
}
\description{
For each intron in the object spongeAnalysis assign unique intron id
}
\examples{

example_dir <- "~/Documents/Projects/15_SpongeAnalysisRpkg/SpongeAnalysis/inst/extdata"
example_files <- fs::dir_ls(example_dir, regexp = "exp*")
names(example_files) <- c("exp1" , "exp2")
x <- read_irfinderS_output(files = example_files,  add_prefix_chr = F)
sponge_analysis_assign_intron_identifier(x)
}
