% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IR_related_functions.R
\name{sponge_analysis_get_introns}
\alias{sponge_analysis_get_introns}
\title{Prepare a master list of introns}
\usage{
sponge_analysis_get_introns(
  f,
  add_meta_data = T,
  add_prefix_chr = FALSE,
  remove_prefix_chr = FALSE,
  bs_genome_object = NULL
)
}
\arguments{
\item{f}{a character string denoting irfinderS output file ending with suffix "IR-nondir".}

\item{add_meta_data}{logical, whether to map meta data or not}

\item{add_prefix_chr}{logical, whether to add prefix 'chr' in the column seqnames}

\item{remove_prefix_chr}{logical, whether to remove prefix 'chr' from the column seqnames}

\item{bs_genome_object}{an object of class BSgenome}
}
\value{
a dataframe
}
\description{
Prepare a master list of introns
}
\examples{
example_dir <- "~/Documents/Projects/15_SpongeAnalysisRpkg/SpongeAnalysis/inst/extdata"
example_files <- fs::dir_ls(example_dir, glob = "*IRFinder-IR-nondir*.txt")
sponge_analysis_get_introns(f = example_files[1],add_prefix_chr = TRUE,bs_genome_object = BSgenome.Hsapiens.UCSC.hg38::BSgenome.Hsapiens.UCSC.hg38)
}
